fixed TypeError in Course class test
modified the tests for the Course class to provide all required arguments when creating a new instance. This resolves the TypeError that was occurring when the tests were run. Each test now provides a specific value for the instance variable it's testing and placeholder values for the other variables

all 3 Course tests now passing