fixed TypeError in Scraper class
modified the Scraper class to create an instance before calling the get_page method. This resolves the TypeError that was occurring when the method was called directly on the class. Also, ensured the Course class is correctly imported from the Course module. This prevents a TypeError when creating a new Course instance