fixed printing of Course instances in Scraper class
changed the way Course instances are printed in the Scraper class. Instead of printing the list of Course instances directly, which calls the __repr__ method of the Course instances, I now loop through the list and print each Course instance individually. This calls the __str__ method of the Course instances, providing a more readable output