added print_courses method to Scraper class and call it at the bottom of the script
Added a print_courses method to the Scraper class to print the details of each Course instance in the courses list. Also, added a call to print_courses at the end of the script to print the courses when the script is run. This allows for easy viewing of the scraped course data